+++ SquiGoWithSwing/src/Game/Model/HUD.java
@@ -4,10 +4,10 @@ import java.awt.*;
 
 public class HUD {
 
-    public static int HEALTH = 100;
-    private int greenValue = 200;
-    private int score=0;
-    private int level=1;
+    public static float HEALTH = 100;
+    private float greenValue = 200;
+    private float score=0;
+    private float level=1;
 
     public void tick() {
         HEALTH = Game.clam(HEALTH, 0, 100);
@@ -19,24 +19,24 @@ public class HUD {
     public void render(Graphics g) {
         g.setColor(Color.gray);
         g.fillRect(15, 15, 200, 32);
-        g.setColor(new Color(100, greenValue, 0));
-        g.fillRect(15, 15, HEALTH * 2, 32);
+        g.setColor(new Color(100, (int)greenValue, 0));
+        g.fillRect(15, 15, (int)HEALTH * 2, 32);
         g.setColor(Color.white);
         g.drawRect(15, 15, 200, 32);
         g.drawString("Score: "+ score,15,64);
         g.drawString("Level: "+ level,15,80);
     }
 
-    public  int getLevel(){
+    public  float getLevel(){
         return  level;
     }
-    public void setLevel(int level){
+    public void setLevel(float level){
         this.level=level;
     }
     public void setScore(int score){
         this.score=score;
     }
-    public int getScore(){
+    public float getScore(){
         return score;
     }
 }
